% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeQuantiles.R
\name{computeQuantiles}
\alias{computeQuantiles}
\title{calculates weighted quantiles}
\usage{
computeQuantiles(xx, ww, qq = 0.5)
}
\arguments{
\item{xx}{un vettore numerico (esempio: salari - MBLS nella RSS)}

\item{ww}{vettore di pesi positivi (esempio: nella RSS, GEWIBGRS)}

\item{qq}{vettore di numeri compresi tra 0 e 1, che indicano il o i livelli di quantile che si vogliono stimare (default 0.5: mediana)}
}
\description{
calculates weighted quantiles
}
