% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\docType{package}
\name{rssci}
\alias{rssci}
\title{Funzioni per la stima di intervalli di confidenza con i dati RSS (salari)}
\description{
In parte riprese dalla documentazione metodologica dell'Ust, 
più correzioni di FC, più eventualmente la stima di intervalli di confidenza
su altri quantili (oltre alla mediana...).
}
\references{
\url{https://www.bfs.admin.ch/bfs/fr/home/statistiques/catalogues-banques-donnees/publications.assetdetail.343582.html}
}
