% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statQuantile_df.R
\name{sep}
\alias{sep}
\title{Estimates the standard error of the empirical cumulative distribution 
function (of wages, for example) at level \code{probs} 
(the quantile level, comprised between 0 and 1)}
\usage{
sep(
  x,
  probs = 0.5,
  strata = NULL,
  psu = NULL,
  nh = NULL,
  th = NULL,
  Nh = NULL,
  mhi = NULL,
  thi = NULL,
  Mhi = NULL,
  weights = NULL,
  crit = NULL
)
}
\arguments{
\item{x}{un vettore numerico (variable di studio di cui si stima la mediana) (esempio: salari (MBLS o MNLIU))}

\item{probs}{livello del quantile desiderato (numero compreso tra 0 e 1)}

\item{strata}{identificatore dello strato}

\item{psu}{identificatore dell'impresa (primary selection unit)}

\item{nh}{numero di imprese sondate nello strato h (dimensioni di campionamento nette degli strati - nel campione originale, indipendentemente da qualsiasi definizione di dominio)}

\item{th}{taux de sondage effectif dans la strate (nh / Nh)}

\item{Nh}{tailles totales des strates}

\item{mhi}{numero di salari sondati nell'impresa i (dimensioni di campionamento delle PSU - nel campione originale, indipendentemente da qualsiasi definizione di dominio)}

\item{thi}{taux de sondage de l'entreprise i (tasso di campionamento delle PSU - mhi / Mhi)}

\item{Mhi}{Dimensioni totali delle PSU}

\item{weights}{pesi delle SSU (tenendo conto dei pesi di campionamento, ma non necessariamente uguali)}

\item{crit}{criterio utilizzato per l'imputazione della varianza per gli strati che contengono una sola PSU}
}
\description{
Estimates the standard error of the empirical cumulative distribution 
function (of wages, for example) at level \code{probs} 
(the quantile level, comprised between 0 and 1)
}
