% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wq2.R
\name{wq2}
\alias{wq2}
\title{Empirical weighted quantile (without ordering)}
\usage{
wq2(x, weights, probs = c(0.5))
}
\arguments{
\item{x}{A numeric vector}

\item{weights}{A vector of (positive) sample weights}

\item{probs}{a numeric vector with the desired quantile levels (default 0.5, the median)}
}
\value{
The weighted quantile (a numeric vector)
}
\description{
This function has to be applied to already sorted data!!!
}
\examples{
wq2(x = rnorm(100), weights = runif(100))
}
\references{
Ferrez, J., Graf, M. (2007). Enquète suisse sur la structure des
 salaires. Programmes R pour l'intervalle de confiance de la médiane. 
 (Rapport de méthodes No. 338-0045). Neuchâtel: Office fédéral de statistique.
}
